refactor(audit): extract occurrence analysis into domain layer
Move the representative-occurrence selection and selector-pattern clustering out
of output/builder/single/findings.rs into a new audit::occurrence_analysis
domain module:

- Result structs RepresentativeOccurrence and FindingPatternCluster move from
  output/report_model.rs; report_model re-exports them so existing references
  keep working.
- Functions build_location_hints, build_representative_occurrences,
  build_pattern_clusters (and their ranking/normalization helpers) move to the
  domain module; findings.rs imports them and stays a pure mapper.

These are locale-neutral heuristics over the domain type OccurrenceDetail, so
the domain is their home. No behaviour change — logic and tests moved verbatim;
dependency direction stays output -> audit.
